refactor: reorganize game components
